fz.itcmdr_clicks
