+++ src/devices/third_reality.ts
@@ -14,6 +14,7 @@ import type {
     Expose,
     Fz,
     KeyValue,
+    KeyValueAny,
     ModernExtend,
     Zh,
 } from "../lib/types";
@@ -189,7 +190,7 @@ interface Third24gRadar {
     commandResponses: never;
 }
 
-const fzLocal = {
+export const fzLocal = {
     thirdreality_acceleration: {
         cluster: "3rVirationSpecialcluster",
         type: ["attributeReport", "readResponse"],
@@ -209,6 +210,16 @@ const fzLocal = {
             return {occupancy: (zoneStatus & 1) > 0};
         },
     } satisfies Fz.Converter<"r3Specialcluster", ThirdMotionSensor, "attributeReport">,
+    itcmdr_clicks: {
+        cluster: "genMultistateInput",
+        type: ["readResponse", "attributeReport"],
+        convert: (model, msg, publish, options, meta) => {
+            const lookup: KeyValueAny = {0: "hold", 1: "single", 2: "double", 3: "triple", 4: "quadruple", 255: "release"};
+            const clicks = msg.data.presentValue;
+            const action = lookup[clicks] ? lookup[clicks] : "many";
+            return {action};
+        },
+    } satisfies Fz.Converter<"genMultistateInput", undefined, ["readResponse", "attributeReport"]>,
 };
 
 export const definitions: DefinitionWithExtend[] = [
@@ -738,7 +749,7 @@ export const definitions: DefinitionWithExtend[] = [
         model: "3RSB22BZ",
         vendor: "Third Reality",
         description: "Smart button",
-        fromZigbee: [fz.itcmdr_clicks],
+        fromZigbee: [fzLocal.itcmdr_clicks],
         ota: true,
         exposes: [e.action(["single", "double", "hold", "release"])],
         extend: [
